feat(wiki): make wiki create/view/edit a routed page
Replace the New-wiki-page dialog and the WikiPageDialog view/edit popup
with a routed page (/skills/wiki/new, /skills/wiki/:slug) mirroring
SkillDetail, for UI consistency with skills. Scope stays governed: it is
freely selectable at create and read-only in edit, with the existing
promotion-request flow for elevation (save_wiki_page takes no scope arg).
Also drop the redundant WikiTab intro paragraph and shorten the two
empty-state descriptions.
